adjust middleware
