+++ src/app/[locale]/layout.tsx
@@ -1,0 +1,53 @@
+import type { Metadata } from "next";
+
+import { cookies } from "next/headers";
+import { getCart } from "@/lib/shopify";
+import { Navbar, Footer } from "@/components";
+import { CartProvider } from "@/contexts";
+import { getMessages, setRequestLocale } from "next-intl/server";
+import { notFound } from "next/navigation";
+import { routing } from "@/i18n/routing";
+import { NextIntlClientProvider } from "next-intl";
+
+export const metadata: Metadata = {
+  title: "Canvas Print Shop",
+  description: "Create your own custom decor",
+  icons: {
+    icon: "/favicon.svg",
+  },
+};
+
+interface Props {
+  children: React.ReactNode;
+  params: Promise<{ locale: string }>;
+}
+
+const RootLayout: React.FC<Props> = async ({ children, params }) => {
+  const cookiesStore = await cookies();
+  const cartId = cookiesStore.get("cartId")?.value;
+  const cart = await getCart(cartId);
+
+  // Ensure that the incoming `locale` is valid
+  const { locale } = await params;
+  if (!routing.locales.includes(locale as any)) {
+    notFound();
+  }
+
+  // Enable static rendering
+  setRequestLocale(locale);
+
+  // Providing all messages to the client
+  // side is the easiest way to get started
+  const messages = await getMessages();
+  return (
+    <NextIntlClientProvider messages={messages}>
+      <CartProvider cart={cart}>
+        <Navbar />
+        {children}
+        <Footer />
+      </CartProvider>
+    </NextIntlClientProvider>
+  );
+};
+
+export default RootLayout;
